initial commit
`scripts-cli/StatementExecutionZIO.scala` is a working ZIO app that runs
under the Scala CLI.  It currently executes a query on a Databricks
SQL Warehouse and waits for the execution to complete using the
Databricks SDK for Java.

It is not currently possible to reliably run ZIO via Ammonite (`amm`)
due to com-lihaoyi/fansi#46.  Fansi chokes when ZIO generates a
warning that contains the ANSI-escape `[2m` (dim/faint mode).